isolate PatcherViewHitTester
